add initialCards to server
