+++ src/components/Section.js
@@ -1,15 +1,12 @@
-import { initialCards } from "../utils/constants.js";
-
 export default class Section {
-  constructor({ items, renderer }, containerSelector) {
-    this._data = items;
+  constructor({ renderer }, containerSelector) {
     this._renderer = renderer;
     this._containerSelector = containerSelector;
     this._container = document.querySelector(containerSelector);
   }
 
-  renderItems() {
-    this._data.forEach((item) => {
+  renderItems(items) {
+    items.forEach((item) => {
       this._renderer(item);
     });
   }
