+++ src/index.html
@@ -53,7 +53,7 @@
       <div class="modal__container">
         <button class="modal__close"></button>
         <p class="modal__heading">Are you sure?</p>
-        <form class="modal__form" id="delete-card-form"></form>
+        <form class="modal__form" id="delete-card-form">
           <button class="modal__button" id="card-close-modal" type="submit">Yes</button>
         </form>
       </div>
