+++ src/utils/constants.js
@@ -33,32 +33,3 @@ export const initialCards = [
     link: "https://www.minnpost.com/wp-content/uploads/2022/03/SouthTemperanceLake940.jpg",
   },
 ];
-
-//FORMS
-export const profileForm = document.forms["profile-form"];
-export const cardAddForm = document.forms["add-card-form"];
-
-// Wrappers
-
-export const profileEditModal = document.querySelector("#profile-edit-modal");
-export const cardAddModal = document.querySelector("#card-add-modal");
-//export const cardsList = document.querySelector(".cards__list");
-
-// Buttons and other DOM nodes
-export const closeButtons = document.querySelectorAll(".modal__close");
-export const profileEditBtn = document.querySelector("#profile-edit-button");
-export const cardAddNewBtn = document.querySelector("#card-add-button");
-
-// Form data
-
-// export const profileTitle = document.querySelector(".profile__title");
-// export const profileDescription = document.querySelector(
-//".profile__description"
-//);
-export const profileInputTitle = document.querySelector("#profile-input-title");
-export const profileInputDescription = document.querySelector(
-  "#profile-input-description"
-);
-
-export const addCardTitleInput = document.querySelector("#card-input-title");
-export const addCardImageInput = document.querySelector("#card-input-url");
